Revert "chore(tests): resolve circular deps related to CVES_DEFAULT_FILTERS and VULNERABILITY_HEADERS"
This reverts commit 756aa334b30e594d081f3809f7b17082ec668d4a.
