+++ src/Components/SmartComponents/CVEs/CVEs.cy.js
@@ -40,7 +40,7 @@ describe('CVE list table', () => {
     it('exists and matches screenshot', () => {
         cy.get('[data-ouia-component-id="cves-table"]');
 
-        cy.get('body').screenshot();
+        cy.get('body').matchImage();
     });
 
     it('has items', () => {
