+++ src/Components/SmartComponents/CVEs/CVEs.test.js
@@ -8,7 +8,7 @@ import configureStore from 'redux-mock-store';
 import { useSelector } from 'react-redux';
 import DownloadReport from '../../../Helpers/DownloadReport';
 import { act } from "react-dom/test-utils";
-import { CVES_DEFAULT_FILTERS } from './CVEsAssets';
+import { CVES_DEFAULT_FILTERS, VULNERABILITIES_HEADER } from '../../../Helpers/constants';
 
 jest.mock('../../../Helpers/Hooks', () => ({
     ...jest.requireActual('../../../Helpers/Hooks'),
@@ -75,6 +75,7 @@ const state = {
         }
     },
     upsupported: false,
+    columns: VULNERABILITIES_HEADER
 };
 
 const customMiddleWare = store => next => action => {
@@ -92,7 +93,7 @@ let wrapper;
 beforeEach(() => {
     store.clearActions();
     useSelector.mockImplementation(callback => {
-        return callback({ CVEsStore: initialState });
+        return callback({ CVEsStore: { ...initialState, columns:VULNERABILITIES_HEADER } });
     });
     wrapper = mountWithIntl(
         <Provider store={store}>
@@ -131,7 +132,8 @@ describe('CVEs', () => {
                 return callback({
                     CVEsStore: {
                         ...initialState,
-                        cveList: { isLoading: false, payload: { errors: true, meta: {}, data: [] } }
+                        cveList: { isLoading: false, payload: { errors: true, meta: {}, data: [] } },
+                        columns: VULNERABILITIES_HEADER
                     }
                 });
             });
