--- src/Components/SmartComponents/CVEs/CVEsAssets.js
@@ -1,58 +1,0 @@
-
-import { classNames, expandable, sortable, nowrap, wrappable } from '@patternfly/react-table';
-import messages from '../../../Messages';
-import { intl } from '../../../Utilities/IntlProvider';
-
-export const CVES_DEFAULT_FILTERS = {
-    affecting: 'true'
-};
-
-export const VULNERABILITIES_HEADER = [
-    {
-        title: intl.formatMessage(messages.cveId),
-        key: 'synopsis',
-        transforms: [sortable],
-        columnTransforms: [wrappable],
-        cellFormatters: [expandable],
-        isShownByDefault: true,
-        isUnhidable: true
-    },
-    {
-        title: intl.formatMessage(messages.publishDate),
-        key: 'public_date',
-        transforms: [sortable, wrappable],
-        columnTransforms: [classNames('hide-description')],
-        isShownByDefault: true
-    },
-    {
-        title: intl.formatMessage(messages.impact),
-        key: 'impact',
-        transforms: [sortable],
-        columnTransforms: [nowrap],
-        isShownByDefault: true
-    },
-    {
-        title: intl.formatMessage(messages.cvssBaseScore),
-        key: 'cvss_score',
-        transforms: [sortable, wrappable],
-        isShownByDefault: true
-    },
-    {
-        title: intl.formatMessage(messages.systemsExposed),
-        key: 'systems_affected',
-        transforms: [sortable, wrappable],
-        isShownByDefault: true
-    },
-    {
-        title: intl.formatMessage(messages.businessRisk),
-        key: 'business_risk',
-        transforms: [sortable, wrappable],
-        isShownByDefault: true
-    },
-    {
-        title: intl.formatMessage(messages.status),
-        key: 'status',
-        transforms: [sortable, wrappable],
-        isShownByDefault: true
-    }
-];
