+++ src/Components/SmartComponents/CVEs/CVEsTable.test.js
@@ -1,6 +1,6 @@
 import CVEsTable from './CVEsTable';
 import toJson from 'enzyme-to-json';
-import { VULNERABILITIES_HEADER } from './CVEsAssets';
+import { VULNERABILITIES_HEADER } from '../../../Helpers/constants';
 import { CVETableContext } from './CVEs';
 import { createCveListByAccount } from '../../../Helpers/VulnerabilityHelper';
 import { handleSortColumn } from '../../../Helpers/MiscHelper';
