+++ src/Components/SmartComponents/CVEs/CVEsTableToolbar.js
@@ -25,11 +25,11 @@ import {
     isFilterInDefaultState
 } from '../../../Helpers/TableToolbarHelper';
 import {
+    CVES_DEFAULT_FILTERS,
     CVES_FILTER_PARAMS,
     ONLY_NON_VULNERABLE_SYSTEMS,
     RULE_PRESENCE_OPTIONS
 } from '../../../Helpers/constants';
-import { CVES_DEFAULT_FILTERS } from './CVEsAssets';
 import { fetchCvesIds } from '../../../Store/Actions/Actions';
 
 const CVEsTableToolbarWithContext = ({ context, canEditStatusOrBusinessRisk, canExport, intl }) => {
