+++ src/Components/SmartComponents/Dashbar/Dashbar.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Card, Grid, GridItem, StackItem, Stack, Alert, CardBody, Text, TextVariants, TextContent } from '@patternfly/react-core';
 import { Main } from '@redhat-cloud-services/frontend-components/Main';
 import { SecurityIcon } from '@patternfly/react-icons';
-import { impactList } from '../../../Helpers/constants';
+import { impactList, CVES_DEFAULT_FILTERS } from '../../../Helpers/constants';
 import { constructFilterParameters, sanitizeLinks } from '../../../Helpers/MiscHelper';
 import { useDispatch, useSelector } from 'react-redux';
 import { changeCveListParameters } from '../../../Store/Actions/Actions';
@@ -14,7 +14,6 @@ import { getAnnouncement, getDashbar } from '../../../Helpers/APIHelper';
 import WithLoader, { LoaderType } from '../../PresentationalComponents/WithLoader/WithLoader';
 import { useIntl } from 'react-intl';
 import sanitizeHtml from 'sanitize-html';
-import { CVES_DEFAULT_FILTERS } from '../CVEs/CVEsAssets';
 
 const DashbarItem = ({ title, count, impact, onLinkClick, hasIcon }) => {
     return (
