+++ src/Helpers/VulnerabilityHelper.test.js
@@ -5,8 +5,8 @@ import {
     handleCVELink,
     renderSystemCSAwLabel
 } from './VulnerabilityHelper';
-import { SYSTEM_DETAILS_HEADER } from './constants';
-import { VULNERABILITIES_HEADER } from '../Components/SmartComponents/CVEs/CVEsAssets';
+import { VULNERABILITIES_HEADER, SYSTEM_DETAILS_HEADER } from './constants';
+
 const { parameters } = initialState;
 
 describe('VulnerabilitiesHelper', () => {
