+++ src/Helpers/constants.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { expandable, sortable, nowrap, wrappable } from '@patternfly/react-table';
+import { classNames, expandable, sortable, nowrap, wrappable } from '@patternfly/react-table';
 import StatusColumn from '../Components/PresentationalComponents/StatusColumn/StatusColumn';
 import AdvisoryColumn from '../Components/PresentationalComponents/AdvisoryColumn/AdvisoryColumn';
 import SystemNameColumn from '../Components/PresentationalComponents/TableColumns/SystemNameColumn';
@@ -500,6 +500,56 @@ export const RISK_OF_CHANGE_LABEL = {
     )
 };
 
+export const VULNERABILITIES_HEADER = [
+    {
+        title: intl.formatMessage(messages.cveId),
+        key: 'synopsis',
+        transforms: [sortable],
+        columnTransforms: [wrappable],
+        cellFormatters: [expandable],
+        isShownByDefault: true,
+        isUnhidable: true
+    },
+    {
+        title: intl.formatMessage(messages.publishDate),
+        key: 'public_date',
+        transforms: [sortable, wrappable],
+        columnTransforms: [classNames('hide-description')],
+        isShownByDefault: true
+    },
+    {
+        title: intl.formatMessage(messages.impact),
+        key: 'impact',
+        transforms: [sortable],
+        columnTransforms: [nowrap],
+        isShownByDefault: true
+    },
+    {
+        title: intl.formatMessage(messages.cvssBaseScore),
+        key: 'cvss_score',
+        transforms: [sortable, wrappable],
+        isShownByDefault: true
+    },
+    {
+        title: intl.formatMessage(messages.systemsExposed),
+        key: 'systems_affected',
+        transforms: [sortable, wrappable],
+        isShownByDefault: true
+    },
+    {
+        title: intl.formatMessage(messages.businessRisk),
+        key: 'business_risk',
+        transforms: [sortable, wrappable],
+        isShownByDefault: true
+    },
+    {
+        title: intl.formatMessage(messages.status),
+        key: 'status',
+        transforms: [sortable, wrappable],
+        isShownByDefault: true
+    }
+];
+
 export const SYSTEMS_EXPOSED_HEADER = [
     {
         key: 'display_name',
@@ -770,6 +820,10 @@ export const SYSTEMS_ALLOWED_PARAMS = [
     'ansible',
     'mssql'
 ];
+
+export const CVES_DEFAULT_FILTERS = {
+    affecting: 'true'
+};
 
 export const SYSTEMS_FILTER_PARAMS = [
     'excluded',
