+++ src/Store/Reducers/CVEsStore.js
@@ -1,6 +1,6 @@
 import * as ActionTypes from '../ActionTypes';
 import { applyGlobalFilter, isTimestampValid } from './reducersHelper';
-import { VULNERABILITIES_HEADER, CVES_DEFAULT_FILTERS } from '../../Components/SmartComponents/CVEs/CVEsAssets';
+import { CVES_DEFAULT_FILTERS, VULNERABILITIES_HEADER } from '../../Helpers/constants';
 
 export const initialState = {
     cveList: {
